Fix annotation for the `parse_dates` argument of the `pandas.read_html` function
In the documentation, it's stated [1] that the usage of the `parse_dates` argument of the `pandas.read_html` function should follow that of the `pandas.read_csv` function. From the documentation of the `parse_dates` argument of the `pandas.read_csv` function [2], it can takes either a "bool or list of int or names or list of lists or dict".

[1] https://pandas.pydata.org/docs/reference/api/pandas.read_html.html#pandas.read_html
[2] https://pandas.pydata.org/docs/reference/api/pandas.read_csv.html#pandas.read_csv