New matplotlib stubs. These were created using a 3-step process:
- using monkeytype to run the matplotlib examples and capture traces
- augmenting the results of that with a parser for the docstrings
- a manual pass to clean up the result

They are still not complete. Anything that had no docstring nor trace
is unlikely to have any annotations unless I spotted something obvious
in the manual pass. And the monkeytype application failed in many cases,
although I did make some changes to make it more robust. The docstring
parses was also not able to handle everything, but did a reasonable job.
Still, these are much more complete than the old stubs.

A couple of things worth noting:

- the docstrings will often call for a "tuple[float, float]" argument;
in practice this need not be a tuple, so I have generally gone with
"Sequence[float]" or sometimes "ArrayLike";
- I believe that the pyplot routines return numpy arrays of Axes, but
I couldn't get that to work; I have gone with lists for now; this probably
needs to be revisited.